JWT/requirements.txt
PyJWT
cryptography
bcrypt


passlib